# apps/api/openapi.yaml
openapi: 3.1.0
info:
  title: Renovix AI API
  description: Unified AI platform with multi-LLM API, chat portal, admin dashboard, RAG knowledge base, AI agents, and MCP integrations.
  version: 1.0.0
  contact:
    name: Renovix AI
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
  - url: https://api.renovix.id/api/v1
    description: Production
security: []
...
paths:
...
  /channels/{id}:
    get:
      tags: [Channels]
      summary: Get channel details
      operationId: getChannel
      security:
        - BearerAuth: []
        - SessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Channel details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Channel"
    patch:
      tags: [Channels]
      summary: Update channel
      operationId: updateChannel
      security:
        - BearerAuth: []
        - SessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateChannelRequest"
      responses:
        "200":
          description: Channel updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Channel"
    delete:
      tags: [Channels]
      summary: Delete channel
      operationId: deleteChannel
      security:
        - BearerAuth: []
        - SessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Channel deleted
...
  /channels/{id}/agents/{agentId}:
    delete:
      tags: [Channels]
      summary: Remove agent from channel
      operationId: removeAgentFromChannel
      security:
        - BearerAuth: []
        - SessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: agentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Agent removed
...
  /agents:
    post:
      tags: [Agents]
      summary: Create an agent
      operationId: createAgent
      security:
        - BearerAuth: []
        - SessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAgentRequest"
      responses:
        "201":
          description: Agent created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
    get:
      tags: [Agents]
      summary: List user agents
      operationId: listAgents
      security:
        - BearerAuth: []
        - SessionAuth: []
      responses:
        "200":
          description: List of agents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Agent"
  /agents/tools/available:
    get:
      tags: [Agents]
      summary: Get available agent tools
      operationId: getAvailableAgentTools
      security:
        - BearerAuth: []
        - SessionAuth: []
      responses:
        "200":
          description: List of available tools
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
...
  /agents/generate-prompt:
    post:
      tags: [Agents]
      summary: Generate agent prompt
      operationId: generateAgentPrompt
      security:
        - BearerAuth: []
        - SessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                description:
                  type: string
      responses:
        "200":
          description: Generated prompt
          content:
            application/json:
              schema:
                type: object
                properties:
                  prompt:
                    type: string
  /agents/clone/{templateId}:
    post:
      tags: [Agents]
      summary: Clone agent from template
      operationId: cloneAgentFromTemplate
      security:
        - BearerAuth: []
        - SessionAuth: []
      parameters:
        - name: templateId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "201":
          description: Agent cloned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
  /agents/subscription:
    get:
      tags: [Agents]
      summary: Get agent subscription
      operationId: getAgentSubscription
      security:
        - BearerAuth: []
        - SessionAuth: []
      responses:
        "200":
          description: Subscription details
  /agents/{id}:
    get:
      tags: [Agents]
      summary: Get agent details
      operationId: getAgent
      security:
        - BearerAuth: []
        - SessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Agent details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
    patch:
      tags: [Agents]
      summary: Update agent
      operationId: updateAgent
      security:
        - BearerAuth: []
        - SessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAgentRequest"
      responses:
        "200":
          description: Agent updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
    delete:
      tags: [Agents]
      summary: Delete agent
      operationId: deleteAgent
      security:
        - BearerAuth: []
        - SessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Agent deleted
  /agents/{id}/status:
    patch:
      tags: [Agents]
      summary: Update agent status
      operationId: updateAgentStatus
      security:
        - BearerAuth: []
        - SessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [active, draft, archived]
      responses:
        "200":
          description: Status updated
  /agents/{id}/publish:
    post:
      tags: [Agents]
      summary: Publish agent
      operationId: publishAgent
      security:
        - BearerAuth: []
        - SessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Agent published
  /agents/{id}/tools:
    post:
      tags: [Agents]
      summary: Attach tool to agent
      operationId: attachAgentTool
      security:
        - BearerAuth: []
        - SessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttachToolRequest"
      responses:
        "201":
          description: Tool attached
  /agents/{id}/tools/{toolId}:
    delete:
      tags: [Agents]
      summary: Remove tool from agent
      operationId: removeAgentTool
      security:
        - BearerAuth: []
        - SessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: toolId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Tool removed
  /agents/{id}/integrations:
    post:
      tags: [Agents]
      summary: Attach integration to agent
      operationId: attachAgentIntegration
      security:
        - BearerAuth: []
        - SessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttachIntegrationRequest"
      responses:
        "201":
          description: Integration attached
  /agents/{id}/integrations/{integrationId}:
    delete:
      tags: [Agents]
      summary: Remove integration from agent
      operationId: removeAgentIntegration
      security:
        - BearerAuth: []
        - SessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: integrationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Integration removed
  /agents/{id}/knowledge-bases:
    post:
      tags: [Agents]
      summary: Attach knowledge base to agent
      operationId: attachAgentKnowledgeBase
      security:
        - BearerAuth: []
        - SessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttachKnowledgeBaseRequest"
      responses:
        "201":
          description: Knowledge base attached
  /agents/{id}/knowledge-bases/{kbId}:
    delete:
      tags: [Agents]
      summary: Remove knowledge base from agent
      operationId: removeAgentKnowledgeBase
      security:
        - BearerAuth: []
        - SessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: kbId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Knowledge base removed
  /agents/{id}/channels:
    post:
      tags: [Agents]
      summary: Deploy agent to channel
      operationId: deployAgentToChannel
      security:
        - BearerAuth: []
        - SessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeployChannelRequest"
      responses:
        "201":
          description: Channel deployed
  /agents/{id}/channels/{channelId}:
    patch:
      tags: [Agents]
      summary: Update agent channel config
      operationId: updateAgentChannel
      security:
        - BearerAuth: []
        - SessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: channelId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Channel config updated
    delete:
      tags: [Agents]
      summary: Remove agent from channel
      operationId: removeAgentFromChannelConfig
      security:
        - BearerAuth: []
        - SessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: channelId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Channel removed
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from /auth/login
    SessionAuth:
      type: http
      scheme: bearer
      bearerFormat: session
      description: Better Auth session token
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: "API key with Bearer prefix: `Bearer sk_live_*`"
    PortalSessionAuth:
      type: apiKey
      in: header
      name: X-Portal-Session
      description: Portal session token for anonymous/free-tier users
...
  schemas:
...
    Channel:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        name:
          type: string
        icon:
          type: string
          nullable: true
        color:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    UpdateChannelRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        icon:
          type: string
        color:
          type: string
...
    Agent:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          nullable: true
        prompt:
          type: string
        status:
          type: string
          enum: [draft, active, archived]
        isPublic:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateAgentRequest:
      type: object
      required: [name, prompt]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
        prompt:
          type: string
        isPublic:
          type: boolean
          default: false
    UpdateAgentRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
        prompt:
          type: string
        isPublic:
          type: boolean
    AttachToolRequest:
      type: object
      required: [toolType]
      properties:
        toolType:
          type: string
        config:
          type: object
        enabled:
          type: boolean
          default: true
    AttachIntegrationRequest:
      type: object
      required: [provider]
      properties:
        provider:
          type: string
        config:
          type: object
        scopes:
          type: array
          items:
            type: string
    AttachKnowledgeBaseRequest:
      type: object
      required: [knowledgeBaseId]
      properties:
        knowledgeBaseId:
          type: string
          format: uuid
    DeployChannelRequest:
      type: object
      required: [channelType]
      properties:
        channelType:
          type: string
        config:
          type: object
...